pyproject.toml
[tool.poetry]
name = "properjpg"
version = "0.3.2"
description = "Make any image ready for the web. Fast."
authors = [
    "Vitor Loureiro <miseravel.cruller-0o@icloud.com>"
    ]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.10",
    "Topic :: Utilities",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Typing :: Typed"
]
license = "MIT"
readme = "README.rst"
homepage = "https://github.com/vitorrloureiro/properjpg"
repository = "https://github.com/vitorrloureiro/properjpg"
documentation = "https://github.com/vitorrloureiro/properjpg"
keywords = ["image", "optimize", "jpg", "conversor", "web"]



[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
pytest-cov = "^3.0.0"
docutils = "^0.18.1"
black = "^22.1.0"
mypy = "^0.931"
types-Pillow = "^9.0.6"
isort = "^5.10.1"

[tool.poetry.scripts]
properjpg = 'properjpg.cli:run'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
